Update user models when new activity happens
