+++ src/server/platform/routes/v1/activity-logging.js
@@ -14,6 +14,7 @@ const cors = require('cors');
 const validator = require('../../../../lib/schema-validator')({
     userActivitySchema: require('../../schemas/user-activity-schema')
 });
+const updateHelper = require('../../../../lib/update-user-models');
 
 // TODO: initialize kafka producer
 // const producer = new KafkaProducer(config.kafka);
@@ -299,6 +300,7 @@ module.exports = function (pg, logger) {
         pg.insert(activity, 'client_activity', (error) => {
             if (error) {
                 // log postgres error
+                console.log(activity);
                 logger.error('error [postgres.insert]: client activity logging failed',
                     logger.formatRequest(req, { error: error.message })
                 );
@@ -309,6 +311,10 @@ module.exports = function (pg, logger) {
                 logger.info('client activity logging successful',
                     logger.formatRequest(req)
                 );
+                // update user model
+                if (activity.uuid != 'unknown:not-tracking'){
+                    updateHelper.updateUserModel(activity);   
+                }
             }
             // send beacon image to user
             return res.sendFile(beaconPath, options);
