+++ src/server/recsys/scripts/load-material-models.js
@@ -7,25 +7,7 @@ const async = require('async');
 // internal modules
 const pg = require('../../../lib/postgresQL')(config.pg);
 const Logger = require('../../../lib/logging-handler')();
-
-function addObjects(objectA, objectB){
-    for (let c in objectB){
-        if (objectA.hasOwnProperty(c)){
-            objectA[c] += objectB[c];
-        }
-        else{
-            objectA[c] = objectB[c];
-        }
-    }
-    return objectA;
-}//function addObjects
-
-function multiplyObjects(objectA, num){
-    for (let c in objectA){
-        objectA[c] *= num;
-    }
-    return objectA;
-}//function multiplyObjects
+const updateHelper = require('../../../lib/update-user-models');
 
 // check if config.schema is defined
 const schema = config.pg.schema;
@@ -175,6 +157,7 @@ function prepareUserModels(callback){
             return process.exit(1);
         }
         if (result.length != 0){
+            console.log('Nothing to update');
             return callback();
         }
         pg.execute(queryUserActivity, [], function(err, result){
@@ -183,83 +166,7 @@ function prepareUserModels(callback){
                 return process.exit(1);
             }
             async.eachSeries(result, function (action, callback){
-                let uuid = action.uuid;
-                let provideruri = action.url;
-                let query = `SELECT * FROM ${schema}.rec_sys_user_model WHERE uuid='${uuid}'`;
-                pg.execute(query, [], function(err, user){
-                    if (err){
-                        console.log('Error fetching user model: ' + err);
-                        return process.exit(1);
-                    }
-                    let escapedUri = provideruri.replace('\'', '\'\'');
-                    let query = `SELECT * FROM ${schema}.rec_sys_material_model WHERE provideruri LIKE 
-                        '${escapedUri}'`;
-                    pg.execute(query, [], function(err, material){
-                        if (err){
-                            console.log('Error fetching material model: ' + err);
-                            console.log('Query: ' + query);
-                            return process.exit(1);
-                        }
-                        if (material.length == 0){
-                            //material is not stored in db
-                            return callback();
-                        }
-                        if (user.length == 0){
-                            user = {
-                                uuid: uuid,
-                                language: {},
-                                visited: {
-                                    count: 0
-                                },
-                                type: {},
-                                concepts: {}
-                            };
-                        }
-                        else user = user[0];
-                        material = material[0];
-                        //check if the user has visited the material before
-                        if (user.visited.hasOwnProperty(material.provideruri)){
-                            // user has already seen the material - nothing to do
-                            user.visited[material.provideruri] += 1;
-                            return callback();
-                        }
-                        //if user has not seen the material
-                        let count = user.visited.count;
-                        let concepts = JSON.parse(JSON.stringify(user.concepts)); // copy concepts object
-                        concepts = multiplyObjects(concepts, count);
-                        concepts = addObjects(concepts, material.concepts);
-                        concepts = multiplyObjects(concepts, 1 / (count + 1));
-                        user.concepts = concepts;
-                        
-                        user.visited[material.provideruri] = 1;
-                        user.visited.count += 1;
-                        
-                        //handle type and language
-                        for (let type in material.type){
-                            if (!user.type.hasOwnProperty(type)){
-                                user.type[type] = 0;
-                            }
-                            user.type[type] += 1;
-                        }
-                        
-                        for (let language in material.language){
-                            if (!user.language.hasOwnProperty(language)){
-                                user.language[language] = 0;
-                            }
-                            user.language[language] += 1;
-                        }
-                        
-                        let conditions = {uuid: uuid};
-                        
-                        pg.upsert(user, conditions, `${schema}.rec_sys_user_model` , function(err){
-                            if (err){
-                                console.log('Error upserting user model: ', + err);
-                                return process.exit(1);
-                            }
-                            callback();
-                        });
-                    });
-                });
+                updateHelper.updateUserModel(action, callback);
             }, function(err){
                  // all jobs are done
                 if (err) {
@@ -283,13 +190,13 @@ function prepareUserModels(callback){
 function initialModelsImport(callback) {
     console.log('Checking whether to update models');
     prepareMaterialModels(function () {
-        //prepareUserModels(function () {
+        prepareUserModels(function () {
             console.log('DONE (initialModelsImport)');
             pg.close();
             if (callback && typeof(callback) === 'function'){
-                    callback();
-                }
-        //});
+                callback();
+            }
+        });
     });
 }//startDbCreate
 exports.initialModelsImport = initialModelsImport;
